Handle invalid search creation
In some scenario subscriber might try to create a new search without
a valid search information, which will raise an :unprocessable error
but our application is not processing that error properly which lead
to some expected behavior.

This commit adds `Search` controller spec to make sure it's behaving
the way it should have and it also properly handle :unprocessable
API response.