Handle invalid search loading
If subscriber tries to load some search that has been expired or they
are not one who created that, then the app does not behave the way it
should have. This commit will handle those invalid search id related
issue and this will also send them back to the search creation page.