Add booking controller specs
In the booking creation the current structure is not doing anything
when the `search_id` or `result_id` is not valid, it just throw the
error page. This commit add controller specs to verify that user is
trying to book a valid search result and it also add that behavior
to the bookings controller.