Add invalid account retrieval spec
If for some reason, the subscriber session has been expired or he
tried to login from another devise then the existing user session
might behave unexpectedly. This commit will handle that exception
and remove the existing session if it has been invalidated so the
user will prompt to re login to the application.