Extract `auth_token` related helpers
In the specs we are using some common helpers related to stubbing the
account authentication, which is lots of duplication and it makes it
harder to mange. This commit extract those helper in a support file
and it also replace/remove the existing helpers we were managing in
each individual file.