Specify `en` as default locale
Some white label are using multiple languages for their landing
page, and when a subscriber is coming from that landing page and
a different locale then the engine is not behaving as expected.

This commit will set the `en` as default, so no matter what it
will always work as expected. Once we have support for other
languages then we can implement some configuration.