Add validations for login model
We do not have any validations for `login`, that's why when ever
user tries to create a new session then we send an API request
regardless the attributes are empty or not.

This commit changes this behavior, and add validations to `login`
model, so no it will ensure, non-empty data are present before
making any actual API request.