Fix postable provider link issue
Some provider requires some specific format and a specific site
id before requesting those protected links, but the site id in
the params is not working as expected.

This commit changes this and use the `site_id` as hard coded
value in the form and that should fix that issue.