Extract company information to configuration
Currently, we are using company address and selling country to some
of the terms and policies related pages, but these values are static
in the engine and it forces us to override those in application.

This commit extract those information to a configuration, so now we
can easily reuse those content by simplify providing those info as
configuration and everything else should work fine.