Add support to create new account
Recently, Discount Network added support to create a new account
in some specific white labels. So, if some white label has been
approved for this then they would like to offer this to the user

This commit take this decision into consideration, so now if a
white label has this option enabled then they should have an
easy access to that feature from this engine.