Specify package id on account creation
The whole subscription process is directly involve with the
packages, and it also define the subscriber's status on how
to charge and when to charge and which is necessary.

This commit adds another configuration option to provide a
`package_id` and it will be used by default if no package is
selected.